requirements.txt
google-genai
docx2python
pandas
GitPython
python-dotenv
pytest